🚚 rename: store 폴더를 common으로 이동 및 가져오기 업데이트 #36
