+++ src/pages/AccountDeletion/components/AccountDeletionForm/index.tsx
@@ -1,11 +1,11 @@
 import Button from "@/common/components/atoms/Button";
 import ReasonSelect from "@/pages/AccountDeletion/components/AccountDeletionForm/ReasonSelect";
-import useAuthStore from "@/stores/authStore";
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import { AccountDeletionFormType } from "@/common/types/formTypes";
 import FormInputField from "@/common/components/molecules/FormInputField";
 import { validationRules } from "@/validationRules";
+import useAuthStore from "@/common/stores/authStore";
 
 export default function AccountDeletionForm() {
   const navigate = useNavigate();
