+++ src/pages/admin/Registration.tsx
@@ -1,13 +1,13 @@
 import { useEffect } from "react";
 import { useForm } from "react-hook-form";
 import Tickets from "../../assets/Tickets.svg";
-import useStore from "../../stores/useStore";
 import { useNavigate } from "react-router-dom";
 import { useLocation } from "react-router-dom";
 import { Match } from "../../common/types/type";
 import axios from "axios";
 import { menu, stadiums, teams } from "../../constants";
 import SuccessToast from "@/common/components/atoms/SuccessToast";
+import useStore from "@/common/stores/useStore";
 
 interface FormValues {
   id?: number; //수정시
